+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/chrome_custom_tabs/ChromeCustomTabsActivitySingleInstance.java
@@ -1,0 +1,7 @@
+package com.pichillilorenzo.flutter_inappwebview.chrome_custom_tabs;
+
+public class ChromeCustomTabsActivitySingleInstance extends ChromeCustomTabsActivity {
+
+  protected static final String LOG_TAG = "ChromeCustomTabsActivitySingleInstance";
+
+}