+++ ios/Classes/InAppWebViewMethodHandler.swift
@@ -290,13 +290,16 @@ public class InAppWebViewMethodHandler: FlutterMethodCallDelegate {
                 break
             case "zoomBy":
                 let zoomFactor = arguments!["zoomFactor"] as! Float
-                let animated = arguments!["iosAnimated"] as! Bool
+                let animated = arguments!["animated"] as! Bool
                 webView?.zoomBy(zoomFactor: zoomFactor, animated: animated)
                 result(true)
                 break
             case "reloadFromOrigin":
                 webView?.reloadFromOrigin()
                 result(true)
+                break
+            case "getOriginalUrl":
+                result(webView?.getOriginalUrl()?.absoluteString)
                 break
             case "getZoomScale":
                 result(webView?.getZoomScale())