+++ lib/src/chrome_safari_browser/android/chrome_custom_tabs_options.dart
@@ -35,6 +35,12 @@ class AndroidChromeCustomTabsOptions
   ///Set to `true` to enable Keep Alive. The default value is `false`.
   bool keepAliveEnabled;
 
+  ///Set to `true` to launch the Android activity in `singleInstance` mode. The default value is `false`.
+  bool singleInstance;
+
+  ///Set to `true` to launch the Android intent with the flag `FLAG_ACTIVITY_NO_HISTORY`. The default value is `false`.
+  bool noHistory;
+
   AndroidChromeCustomTabsOptions(
       {this.addDefaultShareMenuItem = true,
       this.showTitle = true,
@@ -42,7 +48,9 @@ class AndroidChromeCustomTabsOptions
       this.enableUrlBarHiding = false,
       this.instantAppsEnabled = false,
       this.packageName,
-      this.keepAliveEnabled = false});
+      this.keepAliveEnabled = false,
+      this.singleInstance = false,
+      this.noHistory = false});
 
   @override
   Map<String, dynamic> toMap() {
@@ -53,7 +61,9 @@ class AndroidChromeCustomTabsOptions
       "enableUrlBarHiding": enableUrlBarHiding,
       "instantAppsEnabled": instantAppsEnabled,
       "packageName": packageName,
-      "keepAliveEnabled": keepAliveEnabled
+      "keepAliveEnabled": keepAliveEnabled,
+      "singleInstance": singleInstance,
+      "noHistory": noHistory
     };
   }
 
@@ -68,6 +78,8 @@ class AndroidChromeCustomTabsOptions
     options.instantAppsEnabled = map["instantAppsEnabled"];
     options.packageName = map["packageName"];
     options.keepAliveEnabled = map["keepAliveEnabled"];
+    options.singleInstance = map["singleInstance"];
+    options.noHistory = map["noHistory"];
     return options;
   }